Day 09 - Part 01 - Explosives in Cyberspace (Update)
I've updated the solution to only count the expanded characters and
not expand them in the string.  The problem only cares about the
length of the final string, not its contents so there isn't any reason
to expand it.

This will help solving part two as the recursive compression will
probably generate a string larger than memory will allow.

I've also updated the tests to include some for part two.