Day 7 - Part 1 & 2 - Working Solution
It was pretty quick to solve the first part.  The largest amount of
time was parsing the input format into a usable structure.  After that
it was simply building the tree by iterating through the elements and
setting the parent pointer.  Then using any node it is easy to walk up
the parent pointer to find the root and hence the answer.

It took a bit of braining for part two because I didn't have a good
idea how to find the program with the wrong weight.  Since I already
had debug code to dump the tree as a GraphViz dot file I manually
searched the children starting at the root to find the problem node
and then a bit of arithmetic and I had the right answer.

I should solved it by hand earlier.  Doing so gave me an idea for an
algorithm to solve part two.  Implementing the algorithm worked and
didn't take long (< 30 minutes).  This might be a good way to solve
other puzzles I haven't been able to finish from prior years.