Day 08 - Parts 1 & 2 - I Heard You Like Registers
Working solution for part one and two.  Another problem where most of
my time was spent parsing the input data.  Once the data was in a
useable format all that needed to be done was loop through the
instructions and update the appropriate "registers."  I learned a new
use of `std::max_element` which in this case wasn't any shorter than a
standard ranged for loop.  However, using the lambda for the
comparison will be useful in the future when searching a dictionary
for a particular value.

Part two was dead simple.  In the loop just track the maximum value a
register ever reached.  Added one line and returned the both parts in
a tuple.