2017 - Day 12 - Part 1 & 2 - Working Solution
I still had some graph manipulation code from a class I took a few
years ago.  So the hard part was parsing the input to create the graph
and adjacency matrix used to count the components and check for
connectivity between two vertices.  It's slow but it works.

Part 1 future improvement idea: Create a set of each vertice in the
graph.  After making a connection check between vertice 0 and another
vertice increment the connection count and remove every vertice that
has been visited.  This should speed up the algorithm by not checking
for connections we have already found.